Crate forest_runtime

Structs

ConsensusFault

Result of checking two headers for a consensus fault.

Enums

ConsensusFaultType

Consensus fault types in VM.

Traits

ActorCode

Interface for invoking methods on an Actor

MessageInfo

Message information available to the actor about executing message.

Runtime

Runtime is the VM’s internal runtime object. this is everything that is accessible to actors, beyond parameters.

Syscalls

Pure functions implemented as primitives by the runtime.

Functions

compute_unsealed_sector_cid

Computes sector Cid from proof type and pieces for verification.