Trait forest_runtime::ActorCode

pub trait ActorCode {
    fn invoke_method<BS, RT>(
        rt: &mut RT, 
        method: MethodNum, 
        params: &Serialized
    ) -> Result<Serialized, ActorError>
    where
        BS: BlockStore,
        RT: Runtime<BS>;
}

Interface for invoking methods on an Actor

Required methods

fn invoke_method<BS, RT>(
rt: &mut RT,
method: MethodNum,
params: &Serialized
) -> Result<Serialized, ActorError> where
BS: BlockStore,
RT: Runtime<BS>,

Invokes method with runtime on the actor’s code. Method number will match one defined by the Actor, and parameters will be serialized and used in execution

Loading content...