Trait forest_runtime::MessageInfo

pub trait MessageInfo {
    fn caller(&self) -> &Address;

    fn receiver(&self) -> &Address;

    fn value_received(&self) -> &TokenAmount;
}

Message information available to the actor about executing message.

Required methods

fn caller(&self) -> &Address

The address of the immediate calling actor. Always an ID-address.

fn receiver(&self) -> &Address

The address of the actor receiving the message. Always an ID-address.

fn value_received(&self) -> &TokenAmount

The value attached to the message being processed, implicitly added to current_balance() before method invocation.

Loading content...